- Deleted UploadRecord class:   - Another class that was just one method that is only used in Analyster so I moved the method to Analyster and deleted the class.
